modification et ajout fonctions admin
